Add merging functionality of api and rest specs
Added a specfication merger class, responsible for merging rest and api
and added the processing of api.json to the file IO handler.

TODO - test and write test cases
