+++ lib/info_to_spec_mapper.py
@@ -40,8 +40,8 @@ class InfoToSpecMapper():
 
     def get_service_path_from_service_url(self, service_url, base_url):
         if not service_url.startswith(base_url):
-            return service_url
-        return service_url[len(base_url):]
+            return service_url #"/rest" + service_url
+        return service_url[len(base_url):] #"/rest" + service_url[len(base_url):]
 
     def convert_path_list_to_path_map(self, path_list):
         """
